Add IIIFManifest::ManifestRange object
Previously, applications had to roll their own ManifestRange object that
would populate the work's #ranges.  This commit introduces
IIIFManifest::ManifestRange as a first class citizen of the gem.
