+++ README.md
@@ -36,7 +36,7 @@ Additionally it **_may_** implement `#service` to contain an array of hashes for
 
 Additionally it **_may_** implement `#sequence_rendering` to contain an array of hashes for file downloads to be offered at sequences level. Each hash must contain "@id", "format" (mime type) and "label" (eg. `{ "@id" => "download url", "format" => "application/pdf", "label" => "user friendly label" }`).
 
-Finally, it **_may_** implement `ranges`, which returns an array of objects which represent a table of contents or similar structure, each of which responds to `label`, `ranges`, and `file_set_presenters`.
+Finally, it **_may_** implement `ranges`, which returns an array of objects which represent a table of contents or similar structure, each of which responds to `label`, `ranges`, and `file_set_presenters`. `IIIFManifest::ManifestRange` is provided for this purpose; any object with those three methods will also work.
 
 For example:
 
@@ -94,25 +94,16 @@ For example:
 
     def ranges
       [
-        ManifestRange.new(
+        IIIFManifest::ManifestRange.new(
           label: "Table of Contents",
           ranges: [
-            ManifestRange.new(
+            IIIFManifest::ManifestRange.new(
               label: "Chapter 1",
               file_set_presenters: @pages
             )
           ]
         )
       ]
-    end
-  end
-
-  class ManifestRange
-    attr_reader :label, :ranges, :file_set_presenters
-    def initialize(label:, ranges: [], file_set_presenters: [])
-      @label = label
-      @ranges = ranges
-      @file_set_presenters = file_set_presenters
     end
   end
 ```
