+++ lib/iiif_manifest/manifest_range.rb
@@ -1,0 +1,11 @@
+module IIIFManifest
+  class ManifestRange
+    attr_reader :label, :ranges, :file_set_presenters
+
+    def initialize(label:, ranges: [], file_set_presenters: [])
+      @label = label
+      @ranges = ranges
+      @file_set_presenters = file_set_presenters
+    end
+  end
+end
